Added a new usmap package to convert sample area data into a us state map. I extracted the studies that mentioned 2+ states in a single cell, replicated its rows to match the total instances where each state was mentioned, and used rbind to add them all to the Fraud_Location df. I visualized this data into a US State Map.
Determined the total % of mislabeled samples from the total samples and total mislabeled samples.

Determined the instances of confrimed seafood fraud.

Determined the % of seafood labels that match the seafood being sold. Visualized in a bar graph.

Determined top5 seafood groups that get mislabeled.

Used  a dot plot to show the IUCN conservation statuses of the seafood that were given a status.

Used a bar graph to show the common health risks associated with seafood fraud.